+++ cli/src/main/java/edu/uw/apl/tupelo/cli/HashFSCmd.java
@@ -68,11 +68,11 @@ import edu.uw.apl.tupelo.model.Session;
 
 public class HashFSCmd extends Command {
 	HashFSCmd() {
-		super( "hashfs" );//, "Hash file content of a store-managed disk" );
+		// "Hash file content of a store-managed disk" );
+		super( "hashfs" );
 
-		Options os = new Options();
-		os.addOption( "p", false, "print" );
-		options( os );
+		option( "p",
+				"Print existing hashfs result(s) for the identified disk.");
 		requiredArgs( "storeName", "index" );
 	}
 
@@ -83,8 +83,15 @@ public class HashFSCmd extends Command {
 
 		boolean print = cl.hasOption( "p" );
 		String[] args = cl.getArgs();
-		
 		String storeName = args[0];
+		int index = -1;
+		try {
+			index = Integer.parseInt( args[1] );
+		} catch( NumberFormatException nfe ) {
+			HelpCmd.INSTANCE.commandHelp( this );
+			return;
+		}
+
 		Config.Store selectedStore = null;
 		for( Config.Store cs : config.stores() ) {
 			if( cs.getName().equals( storeName ) ) {
@@ -98,46 +105,47 @@ public class HashFSCmd extends Command {
 		}
 		Store store = createStore( selectedStore );	
 
-		int index = Integer.parseInt( args[1] );
-
 		Collection<ManagedDiskDescriptor> mdds = null;
 		try {
 			mdds = store.enumerate();
 		} catch( ConnectException ce ) {
 			System.err.println
 				( "Network Error. Is the remote Tupelo store up?" );
-			System.exit(0);
+			return;
 		}
 		if( index < 1 || index > mdds.size() ) {
 			System.err.println( "Index out-of-range" );
 			return;
 		}
-
-		final boolean debug = true;
 		
 		List<ManagedDiskDescriptor> sorted = new ArrayList( mdds );
 		Collections.sort( sorted,
 						  ManagedDiskDescriptor.DEFAULTCOMPARATOR );
-
 		ManagedDiskDescriptor mdd = sorted.get(index-1);
 
 		if( print ) {
 			report( mdd, store );
 		} else {
-			process( mdd, store );
+			process( mdd, store, storeName );
 		}
 	}
 
-	static void process( ManagedDiskDescriptor mdd, Store s )
+	static void process( ManagedDiskDescriptor mdd, Store store,
+						 String storeName )
 		throws Exception {
 		
+		if( !( store instanceof FilesystemStore ) ) {
+			System.err.println( "Not a FileSystemStore: " + storeName );
+			return;
+		}
+		FilesystemStore fs = (FilesystemStore)store;
+
 		Log log = LogFactory.getLog( HashFSCmd.class );
 		log.info( "Hashing " + mdd );
 		
-		FilesystemStore fs = (FilesystemStore)s;
 		final ManagedDiskFileSystem mdfs = new ManagedDiskFileSystem( fs );
 		
-		final File mountPoint = new File( "mdfs-mount" );
+		final File mountPoint = new File( "mount-hashfs" );
 		if( !mountPoint.exists() ) {
 			mountPoint.mkdirs();
 			mountPoint.deleteOnExit();
